[+AS] Update Gemfile to Ruby 2; update gems & factory
So this app was always supposed to be using ruby 2
but it seems to have been using 1.9.3 for a while.
All the gems have been updated.
The user page spec feature should have been failing
for a while, but wasn't due to something odd with
Capybara 1.x selectors. It is now fixed.
Also fixed a deprecation warning in the project
presenter spec due to upgrading rspec.